+++ libs/third-magic/thirdmagic/utils.py
@@ -44,7 +44,7 @@ def deep_merge(base: dict, updates: dict) -> dict:
 HatchetTaskType = Callable
 try:
     HAS_HATCHET = True
-    from hatchet_sdk.runnables.workflow import BaseWorkflow, Standalone
+    from hatchet_sdk.runnables.workflow import BaseWorkflow, Standalone, Workflow
 
     HatchetTaskType = HatchetTaskType | Callable | Standalone | BaseWorkflow
 except ImportError: